// packages/arknights-schema/src/types/roguelike-topic-table/details/relic-tips-data/relic-tips-data.types.ts
/**

 * JSON 逻辑模块 `details/relic-tips-data/relic-tips-data` 的显式类型。

 */

/**
 * roguelike_topic_table.json 中 RelicTipsData 的显式 TypeScript 定义。
 *
 * 字段语义可在对应中文 Markdown 文档的人工说明区继续补充。
 */
export interface RelicTipsData {
  /** 原始数据字段 `rogue_5_relic_explore_5`。 */
  rogue_5_relic_explore_5?: Rogue5__;
  /** 原始数据字段 `rogue_5_relic_explore_10`。 */
  rogue_5_relic_explore_10?: Rogue5__;
  /** 原始数据字段 `rogue_5_feature_band_28_1`。 */
  rogue_5_feature_band_28_1?: Rogue5__;
  /** 原始数据字段 `rogue_5_feature_band_28_2`。 */
  rogue_5_feature_band_28_2?: Rogue5__;
  /** 原始数据字段 `rogue_5_feature_band_28_3`。 */
  rogue_5_feature_band_28_3?: Rogue5__;
  /** 原始数据字段 `rogue_5_feature_band_28_4`。 */
  rogue_5_feature_band_28_4?: Rogue5__;
}

/**
 * roguelike_topic_table.json 中 Rogue5__ 的显式 TypeScript 定义。
 *
 * 字段语义可在对应中文 Markdown 文档的人工说明区继续补充。
 */
export interface Rogue5__ {
  /** 原始数据字段 `itemId`。 */
  itemId: string;
  /** 原始数据字段 `toastText`。 */
  toastText: string;
}